requirements.txt
pandas==2.3.3
SpeechRecognition==3.15.1
pyttsx3==2.99
groq==1.1.1
python-dotenv==1.2.2
PyAudio==0.2.14
streamlit==1.55.0
streamlit-mic-recorder==0.0.8
requests==2.32.5
gTTS==2.5.4
audio-recorder-streamlit==0.0.10